Move QgsLabelingResults out to its own file
