+++ python/core/core_auto.sip
@@ -327,6 +327,7 @@
 %Include auto_generated/gps/qgsnmeaconnection.sip
 %Include auto_generated/labeling/qgslabeling.sip
 %Include auto_generated/labeling/qgslabelingenginesettings.sip
+%Include auto_generated/labeling/qgslabelingresults.sip
 %Include auto_generated/labeling/qgslabellinesettings.sip
 %Include auto_generated/labeling/qgslabelobstaclesettings.sip
 %Include auto_generated/labeling/qgslabelposition.sip
@@ -629,4 +630,5 @@
 %Include auto_generated/gps/qgsqtlocationconnection.sip
 %Include auto_generated/gps/qgsgpsconnectionregistry.sip
 %Include auto_generated/symbology/qgsmasksymbollayer.sip
+%Include auto_generated/labeling/qgslabelthinningsettings.sip
 %Include auto_generated/qgsuserprofile.sip
