+++ src/app/labeling/qgsmaptoolpinlabels.cpp
@@ -26,6 +26,7 @@
 #include "qgsmaptoolselectutils.h"
 #include "qgsrubberband.h"
 #include "qgslogger.h"
+#include "qgslabelingresults.h"
 
 
 QgsMapToolPinLabels::QgsMapToolPinLabels( QgsMapCanvas *canvas, QgsAdvancedDigitizingDockWidget *cadDock )
